Update File Resource to be smarter about dir creation
Instead of ensuring the directory exists for each file that needs it, we can catalog each dir that needs to
be created, and only try once to create that dir.

We could imrove this more by finding parent dirs that need to be created, and not bother if a child dir is
going to be created, too.