Reverse the order of stack execution for converge, making it top to bottom.
I can't come up with a good reason for it to be either way, but this works with the way I've got the npm stack
written at present, so I'll assume without better information that this is the common use case.

It seems to make sense, actually: setup defaults in the 'parent' stack, and let the stacks that depend
on it overwrite those, which is exactly what I'm doing here.