Make sure resources execute in relative order
Previusly, we only made a resource's usage execution wait until previous executions of itself, but now
we have them waiting for all previous executions of all resources. This ensure that, for instance,
if we want to run the Command resource to execute `npm install`, we know it won't run until after
the  resource finishes executing and writing out the package.json file.